@@ -19,10 +19,8 @@
   },
   "dependencies": {
     "@react-native-community/async-storage": "^1.6.2",
-    "@react-native-community/netinfo": "^4.2.2",
-    "@react-native-firebase/app": "^6.0.0",
-    "@react-native-firebase/messaging": "^6.0.0",
-    "@sentry/react-native": "^1.0.6",
+    "@react-native-community/netinfo": "^4.4.0",
+    "@sentry/react-native": "^1.0.9",
     "apisauce": "^1.1.1",
     "appcenter": "2.4.0",
     "appcenter-analytics": "2.4.0",
@@ -32,22 +30,28 @@
     "moment": "^2.24.0",
     "prop-types": "^15.7.2",
     "react": "16.9.0",
-    "react-native": "0.61.1",
+    "react-native": "0.61.2",
     "react-native-elevated-view": "0.0.6",
     "react-native-fast-image": "7.0.2",
+    "react-native-firebase": "^5.5.6",
     "react-native-gesture-handler": "^1.4.1",
     "react-native-image-gallery": "^2.1.5",
-    "react-native-indicators": "^0.14.0",
+    "react-native-indicators": "^0.17.0",
     "react-native-linear-gradient": "^2.5.6",
-    "react-native-offline": "^5.1.0",
+    "react-native-offline": "^5.2.0",
     "react-native-picker-select": "^6.3.3",
     "react-native-pure-chart": "0.0.24",
+    "react-native-reanimated": "^1.3.0",
+    "react-native-screens": "^2.0.0-alpha.6",
     "react-native-segmented-control-tab": "^3.4.1",
     "react-native-table-component": "^1.2.1",
     "react-native-vector-icons": "^6.6.0",
     "react-native-version-number": "^0.3.6",
     "react-native-version-up": "^1.0.9",
-    "react-navigation": "^3.13.0",
+    "react-navigation": "^4.0.10",
+    "react-navigation-drawer": "2.3.1",
+    "react-navigation-stack": "^1.10.3",
+    "react-navigation-tabs": "^2.5.6",
     "react-redux": "^7.1.1",
     "redux": "^4.0.4",
     "redux-logger": "^3.0.6",
@@ -62,7 +66,7 @@
     "babel-eslint": "^10.0.3",
     "babel-jest": "^24.9.0",
     "babel-plugin-root-import": "^6.4.1",
-    "eslint": "^6.4.0",
+    "eslint": "^6.6.0",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-import-resolver-babel-plugin-root-import": "^1.1.1",
     "eslint-plugin-import": "^2.18.2",
