@@ -1,6 +1,6 @@
 /* eslint-disable implicit-arrow-linebreak */
 /* eslint-disable no-unused-vars */
-import firebase from '@react-native-firebase/app';
+import firebase from 'react-native-firebase';
 
 export default async (message) =>
   // console.log('bgMessaging', message);
